app/schemas/item.py
from pydantic import BaseModel, Field
from typing import Optional

class ItemBase(BaseModel):
    name: str = Field(min_length=1, max_length=200)
    description: Optional[str] = None

class ItemCreate(ItemBase):
    pass

class ItemUpdate(BaseModel):
    name: Optional[str] = Field(None, min_length=1, max_length=200)
    description: Optional[str] = None

class ItemOut(ItemBase):
    id: int
    owner_id: int

    class Config:
        from_attributes = True
