requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic-settings==2.6.0
starlette==0.40.0
starlette-exporter==0.23.0
httpx==0.27.2
pytest==8.3.3
websockets==13.1
loguru==0.7.2

gunicorn==22.0.0

bcrypt==4.0.1

email-validator==2.1.0
